scripts/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
notion-client>=2.2.1
schedule>=1.2.0
